Refactor JSON parsing
